feat: implement get_data_from_tab_layouts
This commit introduces the get_data_from_tab_layouts function.

This function:
- finds the layout of each setup section - eg aerodynamics, transmission etc
- iterates through each layout and finds the widgets
- iterates through the widgets and stores the values of the widgets per section

The function allows me to store the setup values for each area, entered in the slider, as well as the label for which value was selected (e.g. "front wing aero: 4") for each setup section, in a nested dictionary